\encoding{latin1}
\name{CoefVar-package}
\alias{CoefVar-package}
\alias{CoefVar}
\docType{package}
\title{
Paquete CoefVar
}
\description{
El objetivo de este paquete es el de proporcionar una función que calcule el coeficiente de variación de un conjunto de datos.
}
\details{
Puesto que si se calcula el coeficiente de variación usando la función \code{sd} de R lo que realmente se está usando es la cuasi-desviación típica, la única función de este paquete calcula esta medida usando realmente la desviación típica.

Si hubiese más funciones en el paquete, este sería un buen sitio para dar una breve explicación de lo que hace.
}
\author{
Román Salmerón Gómez (\email{romansg@ugr.es})

Maintainer: Román Salmerón Gómez (El que mantiene el proyecto, no tiene por qué coincidir con los autores. Por ejemplo, puede haber más de un autor y que sólo uno de ellos sea el pringado que se encarga de su mantenimiento en CRAn desde el puntos de vista informático).
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{package}
\seealso{
  \code{var}
  \code{sd}
}

