\encoding{latin1}
\name{data.cv}
\alias{data.cv}
\docType{data}
\title{
Datos de prueba
}
\description{
Se trata de un conjunto de datos inventados por mí para ilustrar el funcionamiento de la función que calcula el coeficiente de variación.
}
\usage{data("data.cv")}
\examples{
data(data.cv)
CoeficienteVariacion(data.cv)
}
\keyword{datasets}
